FibonacciHeap
Created the FibonacciHeap and corrected the issue I had with the initial
implementation: when not using the actual CircularDoublyLinkedList data
structure rather, using the techniques directly, have to manually count
the number of nodes in the root list because once nodes are moved from
the root list, we can't tell for sure whether we successfully iterated
through the whole root list.

The MinHeap requires some fixing :(
- The issue is when swapping elements
in the bubbleUp.

Adjusted some properties and values to other data structures including:
- adding the "synchronzied" keyword to methods
- omitting unecessary "final" keywords in properties
- using "protected" for some extended data structures.