Improved Trie
Updated the Trie parseWord() to use the checkWord() because the input
should be parsed to remove any invalid characters before validating

Added the parent and hasWord property as well as the isWord() method on
the TrieNode for the deletion and search processes

Improved the search() process so that is simply checks the hasWord
property to determine whether to continue searching down a word path
because, if the node doesn't have a word or is a word with a value,
then it couldn't contain the word being searched for. This is better
than using the getChild() and checking if the result is null or not.

Improved the structure when removing nodes to perform a trace up the
parents to the root, stopping only when reaching a node that is a word
or contains node(s) with word(s)