TrieHash and Hashtable enumeration
Fixed getPrefix() where the prefix has to be set with a new String of
key + prefix, instead of prefix += key due to the backwards up tracing

Moved the generic methods such as get() and hasWord() into the
AbstractTrie to give a default implementation

Added the isRoot() method on the AbstractTrieNode to check whether the
node is the root or not by checking its key

Added a getRoot() method in the AbstractTrie to circumevent the root
property which won't work with subclass implementations. Has a little
extra overhead due to the call to search() and parseWord(). Allows the
entire Trie walk() and toString() to be default implementations

Added a default constructor in AbstractTrieNode to initialize the root
and hasWord property. The children will still be initialized in
implementing subclasses

Added a constant in the AbstractTrieNode for the root key so it isn't a
magic value when creating a root node or in the isRoot() method

Modified the AbstractEnumerator in the AbstractHashtable to iterate from
the start of the table instead of the end so it is in order and not
reversed

Improved and fixed the delete() method in the Trie where in the up-root
tracing loop conditional statement was node != null, when it should be
parent != null because, the root cannot be removed. Also, added in the
beginning of the loop a quick check if the node is already marked with
not having a child with a word and isn't a word, to immediately remove
it and continue to the next parent node

Added a .gitignore file which currently just ignores any .class file