Refactored and Improved LinkedLists
Made the LinkedLists a single generic type for the entry instead of
having a key/value entry, which is a waste of using a key.

Removed the DoublyLinkedList and instead, made the LinkedList use the
Doubly implementation. The small extra amounts of space requirement
is exceeded by the performance improvement than a basic LinkedList that
can only iterate forwards to find any node or value

Added additional methods such as peek() and poll() as well as the link()
and unlink() methods which help improve code flow and reduce redundancy