Refactored and Improves Tries and Trees, again
Fixed the AbstractTrie hasWord() method. Originally checked if the
returned node wasn't the root. Now it checks if the returned node is a
word

Added a default delete() method in AbstractTrie for a specified node

Made the AbstractTrie iterable since its implementation works for both
the Trie and TrieHash structures

After placing the Enumerator class in the AbstractTrie, did the same for
the AbstractTree since all TreeNodes extend the superclass

Made the EmptyEnumerator class in the default package location so it can
be used instead of the Collections implementation and works for the
Enumerations, Iterators, and Iterables. This reduces the redundancy in
all data structures that can be enumerated

Moved the Entry class from Hashtables to the default package location
since it can be used generally as key/value entity