Prim MST, MinPriorityQueue
Prim MST is not completed and need to verify if the results are correct.

Improved the MinHeap - still needs documenting and revising for a
custom compare function

Created the MinPriorityQueue from the MinHeap but after the minHeapify()
operation in extractMin(), the min value is checked again to ensure if
any changes were made in the queue that the item extracted is the true
minimum

Moved the FiboanacciNode inside the FibonacciHeap as an inner class and
set the contructor private so it can only instantiated within the
FibHeap. The key/value are also private and doesn't require getters for
their values within the FibHeap, making it more practical