Refactored FibonacciHeap
The FibHeap now only inserts the item and the key is delegated to
whatever the compare function is comparing. By default, it will compare
item that is stored via Object.hashCode(). i.e., the Prim algorithm
stores Node instances and uses a custom compare function on the "key"
attribute

Modified the algorithms to import static the Arrays.copyOf method
instead of importing the entire package