Refactored Prim's and MinHeap
Refactored Prim's algorithm to use the new Graph.getVertices(vertex) to
iterate through edges and not all vertices, looking for an edge.
- Removed the visited property on the Prim Node

Fixed the for-loop in Kruskal where the Disjoint-Set was initializing
the vertex as the index value and not the V[i] actual vertex from the
vertices array.

Refactored MinHeap to use the Entry class instead of making a custom one
since it essentially is the same. The main change is in the increaseKey
method where we can't set the "key" attribute, so instead we reassign it
with a new Entry with the new key.