ShorterPathFaster and improvements
Refactored BellmanFord SSSP to use the Graph.getEdges(vertex) method.

Refactored and fixed BFS and DFS with the for-loop
for the vertices to get the actual vertex index from the
G.getVertices() and not using the "i" counter.

More refactoring for static import on Arrays.copyOf instead of importing
the entire Arrays package

Made the Graph checkVertex method public so it can be used by graph
algorithms to validate vertices

Added a has() method in the Queue to determine whether the item exists
in the queue or not.