Major refactoring and improvements
Added the printPath and arrayPath methods to the SSSP class.

Changed the custom exceptions in the Queue, heaps, and Stack to use the
IllegalStateException and NoSuchElementException if the data structure
is adding an element when full or attempting to pull an element when
empty.

Removed the hasNextElement() of the Queue and used the, proper isEmpty()
method.

Fixed the KnapsackMultiple to properly return the total and is left for
refactoring once I figure out how to get item indices.

Refactored FibonacciHeap with more documentation and changed the "n" to
"count" for size counter. Added checks in the extractMin method.

Removed the AbstractQueue class

Renamed the MinHeap minimum() method to peek().