Dijkstra and refactored Graph and path algorithms
Refactored the Graph class to include the Vertex inner class so any
algorithm such as BFS or BellmanFord can easily extend it to hold any
algorithm specific attributes.

The Graph now contains static methods to print or produce an array of a
path with the results of an algorithm.

Refactored the algorithms that produce a path by including calls
to the Graph static methods to return those paths.