Refactored Tries
Removed the seperate Node classes and made them static inner classes.

Upgraded the checkNode() method to be abstract and implemented in any
class that extends the AbstractTrie to verify that the node is an
appropriate instance and not one of another Trie to prevent error.

Set the key for the Hashtable in TrieHash to be the character key
instead of parsing it into an index value, which is unecessary since
the Hashtable essentially will map it for us.

Made the getPrefix() method a default implementation in the AbstractTrie
because it simply retrieves the key value and traces up the parent
property, which is the same for any Trie implementation.