Refactor and improvements for Graph algorithms
Refactored algorithms like DFS, TopologicalSort (DFS), and the SSSP
initSource to include initSourceAll(). This makes the algorithms have
the Node array initialize all vertices, whether or not they exist in the
graph to be able to access the Node[] VTS directly with the vertex as
the index for when 1, 3, 7 exists but not inbetween.

Refactored ShorterPathFaster algorithm in te second for-loop, to first
check the Nodes that if it is unreachable (distance of Infinity), then
to immediately skip and continue to check next vertex to prevent
addition overflow, which could cause it to result in making the "if"
statement true and return null.

Fixed path methods for BellmanFord and ShorterPathFaster, which
returns null for when a cycle is detected.

Fixed the print methods in the SSSP algorithms for directly running and
printing the results which directly used the private _run() method,
which skips the input validation.

Fixed some errors in the documentation for the path methods.

Added the print methods to directly run the algorithm and print results
for SSSP algorithms.