Minimax and refactored FloydWarshall
Adjusted the D and P matrix initialization so that if the graph, W, has
a vertex that doesn't exist in the graph, will fill it with all NIL
values, except for the i == j position which is 0.

Improved space requirements for the algorithm and added a method to
check whether the supplied graph contains a negative weight cycle.