Refactored and improved Trees
Improved documentation

Made AbstractTree and child classes fields private that don't need to be
exposed.

Made the checkType() method abstract to check if the supplied tree node
is null well as validate the node type is used for the correct tree
class.

Made internal main functions for methods like minimum and search, which
recursively calls itself, to only check the arguments once.

Made the tree traversal walk method in the AbstractTree and the actual
_implementation will be implemented in the child class by overriding.