Johnsons, refactored graph algorithms and Queue
Refactored FloydWarshall path methods to first check the results of the
algorithm for a negative weight cycle before passing it to the Graph path
methods.

Refactored graph algorithms to use the for-each loop for the edges.

In the ShorterPathFaster algorithm, checks if the vertex is already in
the queue before adding it to prevent duplicates, since it will be added
only if it was relaxed. It also checks the capacity of the queue
because, if full, that means the graph contains negative weight edges,
which the algorithm doesn't work well with, and could result in an
infinite loop of relaxing the edges, which is another way to detect
negative weight cycles in the algorithm, specifically for this.

Refactored MST algorithms: improved documentation and returns the
Nodes. Extends the Graph.Vertex for the Node class.

Refactored the Queue.enqueue() method to shift the items in the queue if
the tail is at the end but there is space still.

Added the isFull() method to the Queue and Stack to check if the
structure is at the capacity.