Refactored LinkedList and Iterators for classes that implement it
Removed Enumeration implementations because the Iterator should be used
instead.

Renamed the EmptyEnumerator to EmptyIterator since it is an Iterator and
doesn't support Enumerator.

Improved the LinkedList internal Iterator class by making it not generic
to support the items it is holding directly.
Improved the iterator by only holding a reference to the current node
and last node.

Added more methods to LinkedList: clear, removeRange, toArray

Made the LinkedList classes serializable

Slightly refactored the DAGShortestPath algorithm for the iterable
of the LinkedList