Refactored Stack and Queue
Set the Stack and Queue toString method to print in an array format.

Added the iterator for Stack and Queue

Added the modCount increment counter on methods that modify the
structure.

Made the Stack and Queue serializable.

Removed the Enumeration from the LinkedList iterator class.

Improved the toString for the AbstractLinkedList

Updated the iterator and iterable methods for LinkedList since it only
has one type for the elements, there is no need for a generic getter.

Added tests specifically for the iterator remove() method.

Removed the Enumeration from BTree

Updated the EmptyIterator static member

Updated some TODO docs