Refactored BTree
Improved and fixed the walk traversal so that it checks if the node is
null, if not, checks if it is a leaf, if not, calls the callback,
otherwise, it checks each child node and if not null walks through that
child node.

Fixed the iterator remove to not set the lastNode to null because it
can still contain additional keys.