FlowNetwork and refactored Graph
Updated the transpose() to use the for-each loop on the vertices and
edges.

Added the test for cloning a Graph and extending its length.

Moved the Vertex and Edge class higher in the Graph class.

Added a NoSuchElementException throw in getEdgeWeight() if the specified
edge doesn't exist in the graph.

Moved the hasVertex and hasEdge methods above the specific getters to
boilerplating for the checks.

Updated and fixed minor errors in the Graph documentation.

Made the rows, directed, and weighted properties private and made
getters.