Refactored max flow algorithms
Removed the part in the BFS/DFS visits where on each edge, the vertices
were checked to make sure the v wasn't the same as u because of the
reverse edges. It didn't affect the algorithm.

Added the ability for the max flow algorithms to be able to output the
max flow, the string of each flow path and the flow, and a 2-dimensional
array of the paths where the flow for each path is the first value for
each path array.

Refactored the FlowPaths to use LinkedLists to hold all the paths and
parts to build the array paths.