Truly got the log2 now, refactored FibHeap
Found the true log2 function which is log base e of n multiplied by
log base e of 2 (~1.443).

The FibHeap is now serializable.

In the heapLink method of the FibHeap, when inserting y into x's child
list, the child of x must be set to the newly linked child y.

Moved the inner Node class to the top and made private.

Modified the getMin method to return the minimum element without
removing it, rather than exposing the internal node.

Made the decreaseKey method private because of the implementation, only
being able to access the minimum node of the FibHeap, there is no real
practical use for it since the Node is not exposed.

- Left for improvement in the future

Fixed the cut method:

- where it decremented the degree of the parent
node y, even if it might still have another child.

- moved the remove of x from child of y to after checking if x was an
only child. If x was removed before, we can't determine if it was since
it is gone.

- fixed the signature where x is supposed to be the first parameter.

Added the toArray method to the FibHeap.

Refactored the toString method for the FibHeap to return the elements
in an array format.

Fixed the documentation for the toArray.

Removed the util directory under the Trees directory because the java
interfaces are simply imported rather than writing a copy of the basic
functional interface that has no documentation.

Fixed the internal Iterator class for data structures that had missing
documentation.