Refactored MinHeap and minor tweaks to FibHeap
--MinHeap:

Removed the Entry class and instead use the indices as the key.

Added the clear, isFull, toArray, and removeAt methods.

Added the iterator class and is serializable.

Improved the left, right, and parent methods by shifting the bits.

Renamed the minHeapify method to a more proper siftDown.

Removed the increaseKey method.

Using a Comparator for the heck of it to compare values.

--FibHeap:

Added the modCount increments on functions that modify the FibHeap.

Added the clear method to the FibHeap.

Changed the FibHeap getMin method name to peek.