Refactored LinkedLists
Moved the inner Node class to the LinkedList, which involves moving a
lot of methods that use the Node directly into the LinkedList class as
well as making some abstract, and moving the iterator class.

Added the next and prev methods for the Node to be able to traverse the
list from a Node.

Moving the Node to LinkedList was necessary, to allow the nodes to be
referenced so it can continue from the node using the next or prev
methods rather than restarting from the start of end nodes every time
to search for it.