Refactored FlowNetwork and Graph algorithms
When initializing the nodes for the algorithms (VTS), it will create
the array that can hold up to all the possible vertices in the
graph/network, but will only initialize the vertices that exist from
getVertices(). This helps reduce memory usage too because it can be
garbage collected right after the loop exits.

Refactored initSource so that it does as all other algorithms do now
for efficiency: create an array for the total possible vertices but only
initialize the nodes for vertices that do exist.

- Removed the initSourceAll method in SSSP since it is no longer needed
with the improvement.

Refactored the Graph printPathAux and arrayPathAux to work with the
revised SSSP initSource. If the node is null (N[v]), then there is no
possible path from u to v.

Refactored TopologicalSort and DagShortestPath

Refactored ShorterPathFaster

Refactored BFS for the Node to use boolean flags