Refactored RelabelToFront with Queue
Using the Queue to hold the overflowing vertices is much more practical
than original LinkedList implementation. This is because the Queue
removes the overflowing vertex when discharged and adds to the queue any
vertices involved in the discharge operation, rather than removing and
adding a vertex from the list to the front and restarting.

While using the Queue implementation, we can use the Q.size() > 1
condition for the while-loop instead of the !Q.isEmpty() because, the
last vertex that would be dequeued before the algorithm finishes is the
source vertex, which is ineligible to be used for the Discharge
operation.