Refactored Graph, BFS/DFS, FlowNetwork
Removed the getVerticesAll method on Graph and FlowNetwork because their
usage is pretty nonexistent.

Added the getAdjacentVertices method to Graph which is useful for the
BFS/DFS algorithms which don't need the edge weight but rather, just
the adjacent vertices to find paths.

Refactored the algorithms that implement a BFS/DFS algorithms to use the
new Graph.getAdjacentVertices method as long as they only require the
vertex and not any edge property like the weight or flow like in Dinic.