Minor refactoring of RedBlackTree and AVLTree
Changed the color property from an integer to boolean to use less memory
since it is a binary value.

Made the properties of the Node private.

Added checks for the nodes for the color: isRed() and isBlack()

Changed the type from int to boolean for the constants used for
retracing since it is a binary value.