Fixed and refactored BTree
Added parameterized testing for the Many_Insertions class so that it
will now test different minimum degrees "t".

Fixed the delete case 2c, line 1187, where the node key count decrement
wasn't called before shifting and removing the keys.

Fixed the insertNonFull method, where "i" was initialized as the index
for the last key in the node and caused issues when inserting a smaller
key. It now starts at 0 every time and increments up. If a smaller key
is being inserted inbetween keys of a node, it shifts the properly and
inserts the new key in the proper position.

Set the inner BTree node class properties to private so they aren't
exposed.

Added the documentation for the clear() method.