Refactoring Trees - fixes and improvements
Fixed the fencepost problem with the toString methods for the trees.
- Removed the quotes around the results too.

Added the toArray() method to the trees.

Added the modCount increment for the insert and delete methods for the
trees.

Made the BST, AVL, and RedBlackTree insert methods recursive.

Minor error in AVL retracing documentation.

Modfied AVL retracing() to work with the recursive insert
implementation.

Modified the BTreeNode to include keys and values getter since it is
public and can be accessed from the predecessor/succesor Pair methods
as well as the iterator for the entries().

Made the Pair class public and its internal properties private with the
getter methods for the node itself, which has been modified (look above
this) and the key and value from the result of the successor/predecessor
methods.

Made the BTreeNode and Pair classes final.

Refactored the BTree walk() so that it iterates <= node.count, because
the node keys act as a median split between the children nodes keys,
the node will contain node.count + 1 children.

Refactored the toString() method to only iterate up to the node.count,
which is the number of keys, so we know exactly how many keys to iterate
through.