Refactored Tries - major
Because the differences of the normal Trie and TrieHash nodes, they have
their own inner class Node.

The AbstractTrie is now more abstract and less generic default
implementations.

Added the clear method.

Made some fixes to documentation.

Made the search() method a private internal method.

Fixed the insert() method so that it doesn't return the root and will be
able to check the final character with getting the node then checking
the node after. Before, it wouldn't be able to check the final
character node.

Improved the flow of the delete() method.

Refactored the findWords method so that is uses a LinkedList instead of
a Queue to hold the words so that we can manually create the String
array because we cannot cast an Object array to String array.

Made the insert, search, delete, and getPrefix methods recursive.

Made the TrieNode interface to allow some methods to reference the
interface so the nodes implementing it will work.

Made the AbstractNode inner class that needs to be extended in the
implementing Trie class because the normal and TrieHash use different
Nodes and children types.