Refactored Tries - major - cleaner
Removed the AbstractTrie and instead made the Trie fully implemented
with a private inner Node class. The extended TrieHash class operates
the exact same, the only difference being the inner Node class used to
be able to hold the hashtable for the children. It only overrides the
clear and insertRecursive methods, where the inner Node class objects
are instantiated. The constructor also doesn't call super() because we
need to set the root as the TrieHash inner Node and not the default
Node.